refactor: extracted out Routes components ito seperate folder from components
