+++ src/App.js
@@ -1,4 +1,4 @@
-import { RoutesComponent } from "./components/index";
+import { RoutesComponent } from "./Routes/Routes";
 import { Toaster } from 'react-hot-toast';
 import "./App.css";
