+++ src/components/index.js
@@ -1,28 +1,25 @@
-import { Header } from "./Header/Header";
-import { RoutesComponent } from "./Routes/Routes";
-import { CategoryType } from "./Category-type/CategoryType";
-import { ShopNowImgSection } from "./ShopNowImgSection/ShopNowImgSection";
-import { NewArrivalsSection } from "./New-Arrivals/NewArrivalsSection";
-import { DiscountedProducts } from "./Discounted-Products/DiscountedProducts";
-import { ProductsOnSale } from "./Products-on-sale/ProductsOnSale";
-import { Footer } from "./Footer/Footer";
-import { SideBar } from "./Side-bar/SideBar";
-import { ProductsSection } from "./Products-section/ProductsSection";
-import { ProductCardListing } from "./Product-card-listing/ProductCardListing";
-import { PopUpFiltersContainer } from "./PopUp-filters/PopUpFilters";
-import { FilterableForm } from "./Filterable-Form/FilterableForm";
-import { WishlistProduct } from "./Wishlisted-Product/WishlistedProducts";
-import { ProductsInCart } from "./Products-In-Cart/ProductsInCart";
-import { LogInCard } from "./LogIn-card/LogInCard";
-import { LogOutCard } from "./LogOut-card/LogOutCard";
-import { SignUpCard } from "./SignUp-card/SignUpCard";
-import { AprizeLogo } from "./Aprize-logo/AprizeLogo";
-import { CombinedProvider } from "./Combined-providers/CombinedProviders";
-import { FirstVedioCarousel } from "./Carousel/FirstVedioCarousel" ;
-import { MainVedio } from "./Carousel/MainVedio";
-import { Carousel } from "./Carousel/Carousel";
-import { GridTemplate } from "./Grid-template/GridTemplate";
-import { SingleProductDetails } from "./Single-product-detail/SingleProductDetails";
-import { RequiresAuthComp } from "./RequiresAuth/RequiresAuth";
-
-export { Header, RoutesComponent, CategoryType, ShopNowImgSection, NewArrivalsSection, DiscountedProducts, ProductsOnSale, Footer, SideBar, ProductsSection, ProductCardListing, PopUpFiltersContainer, FilterableForm, WishlistProduct, ProductsInCart, LogInCard, LogOutCard, SignUpCard, AprizeLogo, CombinedProvider, FirstVedioCarousel, MainVedio, Carousel, GridTemplate, SingleProductDetails, RequiresAuthComp };
+export { Header } from "./Header/Header";
+export { CategoryType } from "./Category-type/CategoryType";
+export { ShopNowImgSection } from "./ShopNowImgSection/ShopNowImgSection";
+export { NewArrivalsSection } from "./New-Arrivals/NewArrivalsSection";
+export { DiscountedProducts } from "./Discounted-Products/DiscountedProducts";
+export { ProductsOnSale } from "./Products-on-sale/ProductsOnSale";
+export { Footer } from "./Footer/Footer";
+export { SideBar } from "./Side-bar/SideBar";
+export { ProductsSection } from "./Products-section/ProductsSection";
+export { ProductCardListing } from "./Product-card-listing/ProductCardListing";
+export { PopUpFiltersContainer } from "./PopUp-filters/PopUpFilters";
+export { FilterableForm } from "./Filterable-Form/FilterableForm";
+export { WishlistProduct } from "./Wishlisted-Product/WishlistedProducts";
+export { ProductsInCart } from "./Products-In-Cart/ProductsInCart";
+export { LogInCard } from "./LogIn-card/LogInCard";
+export { LogOutCard } from "./LogOut-card/LogOutCard";
+export { SignUpCard } from "./SignUp-card/SignUpCard";
+export { AprizeLogo } from "./Aprize-logo/AprizeLogo";
+export { CombinedProvider } from "./Combined-providers/CombinedProviders";
+export { FirstVedioCarousel } from "./Carousel/FirstVedioCarousel" ;
+export { MainVedio } from "./Carousel/MainVedio";
+export { Carousel } from "./Carousel/Carousel";
+export { GridTemplate } from "./Grid-template/GridTemplate";
+export { SingleProductDetails } from "./Single-product-detail/SingleProductDetails";
+export { RequiresAuthComp } from "./RequiresAuth/RequiresAuth";
